dev: added mockedReduxProvider
