+++ src/UIkit/DappletTextBlock/DappletTextBlock.stories.tsx
@@ -6,13 +6,6 @@ import '/src/index.css'
 
 const meta: Meta<DappletTextBlockProps> = {
   component: DappletTextBlock,
-  decorators: [
-    Story => (
-      <div style={{ width: '400px' }}>
-        <Story />
-      </div>
-    ),
-  ],
   title: 'UIkit/Dapplet Text Block',
   parameters: {
     docs: {
@@ -40,4 +33,11 @@ export const Default: Story = {
     title: 'Title',
     text: 'Semper neque leo scelerisque gravida pharetra, elit viverra varius. Leo et pretium massa tristique mauris habitasse in in. Et cras sociis tellus viverra at dictumst quisque praesent arcu. Eu faucibus id at odio praesent. Ut nibh porta ipsum in eget id netus pharetra in. Ultrices scelerisque augue dui eget lacus, aenean cursus in. Consequat rhoncus egestas ultricies imperdiet diam. Imperdiet mollis egestas est faucibus sit tristique eu.',
   },
+  decorators: [
+    Story => (
+      <div style={{ width: '400px' }}>
+        <Story />
+      </div>
+    ),
+  ],
 }
